Add bit about integrating codelists into SQL
My initial commit didn't show example of integrating codelists into a SQL statement...which is a big oversight because that was the main point of the whole notebook, in the first place!

I've also uploaded the example codelist and the images that are displayed.

Finally, I added a wee explained of codelists, based on Megan's feedback.

Let me know what you think @JohnDomar @ShoreRob1 @DrMeganLWood