Remove Developmental Academic Disorder from caseness definition
In a meeting between CMI and CB on 10th August 2023, CB decided that Developmental Academic Disorder should not be a criterion diagnosis for the caseness of complex mental health difficulties. In an email sent on the 9th August, CB said:

> "I don't think Developmental Academic Disorder should be in our caseness set.
I don't recall it as being in our conceptualisation (some neurodev disorders such as ASDs are in the complex feature sets but shouldn't be cases).
> What's more I  stumbled upon the term today in practice - it's not one I recognise from UK clinical practice or Read codes but it's essentially the root of learning disability (I did a LD review on someone and saw these SNOMED codes appearing!).   I have checked SNOMED CT and the hierarchy goes:
? Mental Disorders / Developmental Mental Disorders / Developmental Academic Disorders / {Learning disability etc.....}

> We shouldn't have DADs in the mix. I suspect that quite a lot of these will be in single-code "cases" so it will probably bring the caseness prevalence down quite a bit further.  "