Update to H.M.A. calculation
When calculating whether or not a patient record was indicating an Adjust state or not, the beginning of their observations would be Adjust by default because there was no prior prescription. I've change the syntax such that I look backward from the start of the observation window to figure out if the first-observed intertest interval is truly an Adjust or not.