Understand issue with non-narrowing plots
I was concerned that the non-narrowing plots were somehow carrying forward dropouts. I then realised that the plots are of data from a cohort of records with at least as many events as is being plotted. In other words, the drop outs were already removed so no narrowing would be seen in any plots.